+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - alias: add ability to use a different netowrk interface for alias network type (#80)
+- env: add pot info -E output to environment or tinirc
 
 ### Fixed
 - prepare: fix multiple export port support